+++ company/src/main/java/com/hubEleven/company/infrastructure/client/SlackClient.java
@@ -1,6 +1,6 @@
 package com.hubEleven.company.infrastructure.client;
 
-import com.hubEleven.company.infrastructure.configuration.FeignConfig;
+import com.hubEleven.company.infrastructure.config.FeignConfig;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
