+++ company/src/main/java/com/hubEleven/company/infrastructure/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package com.hubEleven.company.infrastructure.configuration;
+package com.hubEleven.company.infrastructure.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
